Added drafts of adapter classes to JavaFX FileChooser
Idea is to have at least some kind of
modality of JavaFX FileChooser in Swing.
Neither FileChooser or DirectoryChooser
provide proper implemented modality
when used in Swing.

The drafted adapter now at least ensured,
that the root Swing panel or SwingNode
is blocked so that not more than one dialog
can be opened at once.

Nevertheless, the JavaFX director and file
chooser instances remain accessible and
can be configured as liked.